feat - split learn and remove main
