+++ run.py
@@ -1,15 +1,18 @@
-from models.q_learning.qr_dqn.main import qrdqn_main
+from enum import Enum
+import gym
+from models.q_learning.qr_dqn.agent import Agent
+from variables.gamename import GameName
+from variables.hyperparams.q_learning.qr_dqn import HyperParams
 
 
-game_name = "CartPole-v1"
-params_dict = {
-    "gamma": 0.99,
-    "batch_size": 8,
-    "lr": 1e-4,
-    "atoms": 8,
-}
+def train(
+    game_name: str,
+    hyper_params: Enum,
+):
+    env = gym.make(game_name)
+    agent = Agent(env, hyper_params)
+    agent.train()
 
-qrdqn_main(
-    game_name=game_name,
-    params_dict=params_dict,
-)
+
+if __name__ == "__main__":
+    train(game_name=GameName.CARTPOLE_V1.value, hyper_params=HyperParams)
